论坛最基本的功能接口

+++ src/main/java/org/applesline/forum/model/Comments.java
@@ -15,15 +15,16 @@ import org.beetl.sql.core.annotatoin.Table;
 @Table(name="t_comments")
 public class Comments {
 
-	private Integer id;
+	private long id;
 	private long articleId;
-	private Integer userId;
+	private long userId;
 	private String content;
 	private Date createTime;
-	public Integer getId() {
+	
+	public long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public long getArticleId() {
@@ -32,13 +33,12 @@ public class Comments {
 	public void setArticleId(long articleId) {
 		this.articleId = articleId;
 	}
-	public Integer getUserId() {
+	public long getUserId() {
 		return userId;
 	}
-	public void setUserId(Integer userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
-
 	public String getContent() {
 		return content;
 	}
